# Vectorize `find_first_of` for 4 and 8 byte elements