… benchmarks/src/find_first_of.cpp