@AlexGuteniev 2024-04-14
> Looks like we will have way better results if specialized approach is generated, instead of having different and simpler generic approach. But the complexity would grow.

On the other hand, despite having more code, it would be an uniform approach.